AI_Manager/requirements.txt
azure-core==1.23.1
azure-identity==1.9.0
azure-storage-file-share==12.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.2
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
MarkupSafe==2.1.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
portalocker==2.4.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sklearn == 0.0
urllib3==1.26.9
zipp==3.8.0
kafka-python==2.0.2
dill == 0.3.4
opencv_python == 4.5.5.64
keras == 2.8.0
matplotlib == 3.4.3
absl-py==1.0.0
astunparse==1.6.3
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
deepface==0.0.74
fastjsonschema==2.15.3
filelock==3.6.0
flatbuffers==2.0
gast==0.5.3
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
mtcnn==0.1.1
nbformat==5.3.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
pandas==1.4.2
pigar==0.10.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.10
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
imutils==0.5.4
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
deepface==0.0.74
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.2
flatbuffers==2.0
gast==0.5.3
gdown==4.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.4.0
jupyter-core==4.10.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
mtcnn==0.1.1
nbformat==5.3.0
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
pandas==1.4.2
pigar==0.10.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
retina-face==0.0.10
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
azure-core==1.23.1
azure-storage-file-share==12.7.0
